requirements.txt
Flask==3.0.3
Jinja2==3.1.4
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1


Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
